feat: create components for left and right side in hero section
